uv/tests: make `python_version < '0'` regression test smaller
Running this test manually on the latest release shows that it still
emits a `python_version < '0'` marker:

    $ uv-0.4.29 pip compile requirements.in -c constraints.txt --annotation-style line --python 3.10 --universal | rg "< '0'"
    Resolved 151 packages in 97ms
    apache-airflow-providers-common-sql==1.19.0 ; python_version < '0'  # via apache-airflow-providers-sqlite

That is, even though this is a smaller test, it's still testing the same
bug.

This is about twice as fast on my machine. It's probably still worth
moving to a packse test, but this was quick to do.