uv-resolver: refactor forking logic
This is another refactor of the forking logic that essentially
abstracts out `MarkerTree`. That is, instead of doing forking in a
way that is tightly coupled with `MarkerTree`, we use an intermediate
abstraction that hides the `MarkerTree`. Internally, it is still just
using a `MarkerTree`, but the intent is to expand this with conflicting
groups in a subsequent commit.

The main idea here is that `Fork` now has a `ResolverEnvironment`
instead of a `MarkerTree`. And a somewhat sneaky simplification here
is that forking now takes the parent `ResolverEnvironment` and starts
its forking logic from there. I believe we used to do this (or wanted
to do this) in the past, but stopped because of our sub-optimal marker
simplification. But with that fixed, we can do things this way now
which is much simpler than trying to combine them later.

(I'm starting to suspect that `PythonRequirement` should also live in
`ResolverEnvironment` too, but I'm holding off on further refactoring.)