uv-resolver: thread conflicting groups into the lock file
We need to write conflicting groups to the lock file so that
we can check for changes to it compared to the last resolution.

This is similar to how we deal with "supported environments."
When a change happens, it can impact resolution, so we need to
force a re-resolve.

Moreover, specific to conflicting extras, we need to report
an error at install time if conflicting extras are requested.