uv-workspace: make package name option in conflicting-groups schema
In order to support this, we define a nearly duplicative set of types
just for `pyproject.toml` deserialization. We also permit the conflicts
to be specified in workspace member `pyproject.toml` files. So now we
collect all of them and merged them into one giant set to feed to the
resolver. When a package name isn't specified, it is implicitly assumed
to be the name of the project that defined the conflicts.

Note that this also removes support for specifying conflicts in
`uv.toml`. I think I didn't mean to add that originally. We could add it
back, but in that context, we would need the package name I believe.