uv-resolver: error on unconditional conflicting extras
If we don't do this, then it would be like permitting
`uv sync --extra x1 --extra x2` even when `x1` and `x2` are declared as
conflicting.

Technically, we should only report an error when 2 or more conflicting
extras are unconditionally enabled. Instead, here, we report an error
if just 1 is found. The reason is that it seems tricky to detect all
possible cases of 2 or more since I believe it would require looking at
the full dependency tree.