Treat less compatible tags as lower priority in resolver (#9339)
## Summary

This is a second pass at https://github.com/astral-sh/uv/pull/7556,
which was reverted in https://github.com/astral-sh/uv/pull/7608 due to a
regression in https://github.com/astral-sh/uv/issues/7606. The behavior
is actually correct, but a package (`nmslib`) publishes inconsistent
metadata, and the change here happened to cause us to select a wheel
with "wrong" metadata. It's arbitrary, but it did cause a regression for
folks.

Since we're now seeing other issues caused by the wrongness here (and
since the reporter in https://github.com/astral-sh/uv/issues/7606 has
since removed the dependency), I'm inclined to ship this fix.

Closes https://github.com/astral-sh/uv/issues/7553.
Closes https://github.com/astral-sh/uv/issues/9283.